Created parent Runner class and moved rainbow runner and colour runner into one file
