+++ software/thegrid/patterns/__init__.py
@@ -23,7 +23,7 @@ from . import radar
 from . import colourwave
 from . import grid_of_life
 from . import colour_ripple
-from . import colour_runner
+from . import runner
 from . import stopping
 from . import soundtest
